requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
ConfigArgParse==1.5.3
DateTime==4.4
Deprecated==1.2.13
et-xmlfile==1.1.0
execnet==1.9.0
Flask==2.1.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.0.3
jsonpath==0.82
locust==2.8.6
loguru==0.6.0
MarkupSafe==2.1.1
mock==4.0.3
msgpack==1.0.3
Naked==0.1.31
openpyxl==3.0.9
packaging==21.3
playwright==1.21.0
pluggy==1.0.0
psutil==5.9.0
py==1.11.0
pycrypto==2.6.1
pyee==8.1.0
PyMySQL==1.0.2
pyparsing==3.0.8
pytest==7.1.1
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
redis==4.2.2
requests==2.27.1
roundrobin==0.0.2
sha256==0.3
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
websockets==10.1
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
